requirements.txt
wxpy
qiniu
beautifulsoup4
lxml
sqlalchemy
Flask-SQLAlchemy
requests